# backend/src/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "FileBrowser Backend API",
    "description": "Full API reference and interactive testing interface for the FileBrowser Node.js/Express backend.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/api",
      "description": "API Base URL"
    }
  ],
  "security": [
    {
      "BearerAuth": []
    },
    {
      "XAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter the JWT token (without 'Bearer ' prefix)."
      },
      "XAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Auth",
        "description": "Alternative authorization header using 'X-Auth'."
      }
    },
    "schemas": {
...
      "Share": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "hash": { "type": "string" },
          "path": { "type": "string" },
          "userId": { "type": "integer" },
          "expire": { "type": "integer" },
          "createdAt": { "type": "integer" }
        }
      },
...
    }
  },
  "paths": {
...
    "/search": {
      "get": {
        "tags": ["Search"],
        "summary": "Search root folder",
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "required": true,
            "schema": { "type": "string" },
            "description": "Term or glob pattern (e.g. *.pdf, type:image)"
          }
        ],
        "responses": {
          "200": {
            "description": "NDJSON stream of files",
            "content": {
              "application/x-ndjson": {
                "schema": {
                  "type": "string",
                  "example": "{\"path\":\"/image.png\",\"dir\":false}\n{\"path\":\"/photo.jpg\",\"dir\":false}"
                }
              }
            }
          }
        }
      }
    },
    "/search/{path}": {
      "get": {
        "tags": ["Search"],
        "summary": "Search under path",
        "parameters": [
          {
            "name": "path",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "name": "query",
            "in": "query",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "NDJSON stream of files"
          }
        }
      }
    },
...
    "/share/{path}": {
      "get": {
        "tags": ["Shares"],
        "summary": "List share links for specific path",
        "parameters": [
          {
            "name": "path",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Share" }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Shares"],
        "summary": "Create share link for path",
        "parameters": [
          {
            "name": "path",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "expires": { "type": "string", "example": "7" },
                  "unit": { "type": "string", "enum": ["seconds", "minutes", "hours", "days"], "example": "days" },
                  "password": { "type": "string", "example": "secret123" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Share" }
              }
            }
          }
        }
      }
    },
    "/share/{hash}": {
      "delete": {
        "tags": ["Shares"],
        "summary": "Delete a share link",
        "parameters": [
          {
            "name": "hash",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "Deleted successfully" }
        }
      }
    },
    "/public/share/{hash}": {
      "get": {
        "tags": ["Shares"],
        "summary": "Public: get share link details",
        "security": [],
        "parameters": [
          {
            "name": "hash",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "name": "password",
            "in": "query",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Public share file info"
          }
        }
      }
    },
    "/public/dl/{hash}": {
      "get": {
        "tags": ["Shares"],
        "summary": "Public: download shared file/directory",
        "security": [],
        "parameters": [
          {
            "name": "hash",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "name": "password",
            "in": "query",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "File stream or zip stream download" }
        }
      }
    },
...
    "/tus/{id}": {
      "head": {
        "tags": ["TUS"],
        "summary": "TUS check offset",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "Upload info, upload-offset returned" }
        }
      },
      "patch": {
        "tags": ["TUS"],
        "summary": "TUS append chunk",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "name": "Upload-Offset",
            "in": "header",
            "required": true,
            "schema": { "type": "integer" }
          },
          {
            "name": "Content-Type",
            "in": "header",
            "required": true,
            "schema": { "type": "string", "example": "application/offset+octet-stream" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/offset+octet-stream": {
              "schema": { "type": "string", "format": "binary" }
            }
          }
        },
        "responses": {
          "204": { "description": "Chunk processed, upload-offset returned" }
        }
      },
      "delete": {
        "tags": ["TUS"],
        "summary": "TUS terminate upload",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "204": { "description": "Terminated successfully" }
        }
      }
    }
  }
}